FM1D2D-403 Fixed several Structure File reading/writing related issues for both 1D and 2D, including: - updating property names - adding missing properties to GUI - fixing cross section definition files not being created and several other issues related to cross section definition in the context of structures. - fixing several exceptions that occured - making sure all properties are correctly written to the structures.ini - fixing 1D structures not being loaded when loading a project
Former-commit-id: 9758be8ef955f3bf42af51bf7e32247ba25bc329
Former-commit-id: b670e76c14f6d9560cf270d1394095809073cfd9

+++ src/DeltaShell.NGHS.IO/FileReaders/Definition/Structures/WeirDefinitionReader.cs
@@ -87,15 +87,15 @@ namespace DeltaShell.NGHS.IO.FileReaders.Definition.Structures
                     return new PierWeirFormula
                     {
                         NumberOfPiers = category.ReadProperty<int>(StructureRegion.NPiers.Key),
-                        UpstreamFacePos = category.ReadProperty<int>(StructureRegion.PosHeight.Key),
-                        DesignHeadPos = category.ReadProperty<int>(StructureRegion.PosDesignHead.Key),
-                        PierContractionPos = category.ReadProperty<int>(StructureRegion.PosPierContractCoef.Key),
-                        AbutmentContractionPos = category.ReadProperty<int>(StructureRegion.PosAbutContractCoef.Key),
+                        UpstreamFacePos = category.ReadProperty<double>(StructureRegion.PosHeight.Key),
+                        DesignHeadPos = category.ReadProperty<double>(StructureRegion.PosDesignHead.Key),
+                        PierContractionPos = category.ReadProperty<double>(StructureRegion.PosPierContractCoef.Key),
+                        AbutmentContractionPos = category.ReadProperty<double>(StructureRegion.PosAbutContractCoef.Key),
 
-                        UpstreamFaceNeg = category.ReadProperty<int>(StructureRegion.NegHeight.Key),
-                        DesignHeadNeg = category.ReadProperty<int>(StructureRegion.NegDesignHead.Key),
-                        PierContractionNeg = category.ReadProperty<int>(StructureRegion.NegPierContractCoef.Key),
-                        AbutmentContractionNeg = category.ReadProperty<int>(StructureRegion.NegAbutContractCoef.Key)
+                        UpstreamFaceNeg = category.ReadProperty<double>(StructureRegion.NegHeight.Key),
+                        DesignHeadNeg = category.ReadProperty<double>(StructureRegion.NegDesignHead.Key),
+                        PierContractionNeg = category.ReadProperty<double>(StructureRegion.NegPierContractCoef.Key),
+                        AbutmentContractionNeg = category.ReadProperty<double>(StructureRegion.NegAbutContractCoef.Key)
                     };
                 case StructureType.Orifice:
                     return new GatedWeirFormula
